requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-filter>=23.1
dj-database-url>=1.0
gunicorn>=20.1
whitenoise>=6.5
psycopg2-binary>=2.9
django-cors-headers>=4.0
django-redis>=5.2
redis>=4.0
Pillow>=9.0

# Optional / dev packages (not strictly required in production but useful)
pytest>=7.0
pytest-django>=4.5

# Production / deployment helpers
django-jazzmin>=2.0
mysqlclient>=2.1
python-dotenv>=1.0